scripts/upload_test_results/requirements.txt
certifi==2020.6.20
configparser==5.0.0
gitdb==4.0.5
GitPython==3.1.7
smmap==3.0.4
wget==3.2
